🦄 refactor(orm): update for sea-orm
use sea-orm replace rbatis

+++ src/models/vo/serise.rs
@@ -1,13 +1,13 @@
 use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash, Default)]
 pub struct Series {
-    id: u16,
+    id: i64,
     name: String,
-    value: i32,
+    value: u64,
 }
 
 impl Series {
-    pub fn new(id: u16, name: String, value: i32) -> Self {
+    pub fn new(id: i64, name: String, value: u64) -> Self {
         Self { id, name, value }
     }
 }
